chore: separate out converters
